1. Dihedral angle potential added (computes both proper and improper dihedrals). To enable:
dihedral                on
dihedral_proper_func    1
dihedral_improper_func  2

Potential is disabled by default. Function types 1 for proper and 2 for improper are default. Dihedrals and their parameters are read from the corresponding section of the topology file with Gromacs units and order of the parameters (see Table 5.5 of Gromacs Manual).

2. Clean-up: old read_topology, read_parameters and ComputationalArrays files are moved to Trash folder since they are not in use anymore.

3. Small fix to AngleHarmonic Potential (sizeof(int) -> sizeof(float)).

+++ src/Potentials/AngleHarmonic.cu
@@ -91,7 +91,7 @@ AngleHarmonic::AngleHarmonic(MDData *mdd, int angleCount, int3* angles, float2* 
 	cudaMemcpy(d_pars, h_pars, angleCount*sizeof(float2), cudaMemcpyHostToDevice);
 	cudaMemcpy(d_refs, h_refs, angleCount*sizeof(int4), cudaMemcpyHostToDevice);
 	cudaMemcpy(d_count, h_count, (lastAngled + 1)*sizeof(int), cudaMemcpyHostToDevice);
-	cudaMemcpy(d_energies, h_energies, angleCount*sizeof(int), cudaMemcpyHostToDevice);
+	cudaMemcpy(d_energies, h_energies, angleCount*sizeof(float), cudaMemcpyHostToDevice);
 
 	printf("Done initializing Harmonic Angle potential\n");
 }
